Get basic Jest testing setup.
This commit replaces razzle test with calling jest directly.
This means that all the config razzle generates for jest needed to be
replicated, and saved in the package.json.

Currently the installed version of jest is locked in at 20.0.4 to match
the one that Razzle uses. Without this we get "TypeError:
environment.setup is not a function".