Delete svg, remove whitespace from markup
The svg was there as a potential 404 landing image.
In the end, we decided not to use it, so we don't need it.

Whitespace between the tag name and then content
- <H1> Title </H1>
vs
- <H1>Title</H1>

is ignored by browsers when HTML is rendered, and it results in
worse-looking formatting in our components.

Removing it makes everything better.