Move global styles to Layout
Also change the layout element that styling passed into Layout is
applied to.
When it was being applied to the <main> element, I was having to
write more specific CSS that I really wanted.
Now we have less side-effects because our CSS only applies to
markup that our page components know about.