Integrate Calendar with react-final-form
react-final-form is the library we are using for the forms on the
registration page. We want to keep using this library as much as
we can (for consistency reasons, etc.)

I've got it working such that the form can see the selected dates
at the time the form is submitted, but our Calendar component is
_not_ able to be validated inline because the Form's validate
method appears to only be triggered when Calendar is focused or
blurred, not when additional days are clicked.

That is, when the Calendar is first clicked, the validate method is
run once. When additional days are clicked, the validate method is
_not_ run.
By contrast, the validate method is run on each change to the native
text `<input>` element.

It seems like the problem is that `<input>`s have a default `onChange`
attribute that the form library expects, whereas our Calendar is
just a `<div>` wrapping more `<div>`s, and it doesn't have a
native `onChange` attribute.

Currently, the Calendar element is working with `react-final-form`
enough that we can move forward. Its values are bubbled up to the
form, but only onSubmit, not on validate.

We don't need inline validation for our purposes, so this isn't a
showstopper for us, but it would be good to solve this problem
because it feels like it will trip us up eventually.

Here is an issue I opened on the React Day Picker repo:
https://github.com/gpbl/react-day-picker/issues/705

It explains the problem more clearly and has an interactive example.