Add FieldAdapter PropTypes to our subclassed input fields
The other option is possibly keeping all the Adapter components
together in their own file, but for now let's do it this way.

In addition to this, I needed to add an (empty) onChange function
to the tests for all of our Adapter components because before I
hadn't been specifying that `onChange` was required but now I am.