Remove statefulness from tests
Accidentally, I was passing around values between tests that
affected the other tests.

Yikes.

Also, as a follow-up to the idea that we should build a Calendar
Adaptor, I was looking at doing that but it seemed like it would
be tricky, and we don't actually need our Calendar component
anywhere outside of a form.
So I'm going to forget that for now.