Add tabIndex to language-switcher link
Since our language switcher is a link with no `href` attribute, it
was being left of of the default tab order (ie, there was no way
to navigate to it with a keyboard).

Giving it a tabIndex means that it can be focused on, and I also
added a keyPress listener that activates it if the enter key or
the spacebar is pressed.

In its current form (responding to spacebars and enters), it should
just be a <button> element, but I think we're planning to actually
link to something so we'll leave it as a link for now.