Turn sections into divs
Sections are supposed to have headers, which implies that they
should be about breaking up content on the page.

From MDN:

> Do not use the <section> element as a generic container;
> this is what <div> is for

For our Footer and AlphaBanner and Content, we're basically
just using the <section> element as a container,
so I made them divs.

This should also silence some of our HTML validation warnings.