Save selectedDays
Extend the CalendarPage to save the selectedDays
into the ApolloLinkState when the submit button is pressed.

** Also **

Change string representation of dates
- We want humans to read them
- We don't care about the 'time' part of this

Start to standardize dateToString utilities

We're going to want to create (at least) a couple different ways
to standardize how we transform/show Dates:
- a friendly Date string that we will show to users
- an ISO Date string that the <time> element needs

Technically the DayPicker is returning `Date()` objects,
but we're not really allowing people to pick times so
we are only keeping the year, month, and day in our string
representations.