Simplify returning values to registration form
I was poking around in here to see how to repopulate saved values
to the form (because I have to do the same for the calendar page),
and it looked like the GET call (to see what our saved form values
are) was being duplicated.

I've cut out some of the duplication so that our code is simpler.