Return previously selected dates to Calendar
When you 'go back' from the review page, you'll get your selected
days back. That's sweet.

Worth noting that the `makeDate` method needs to convert strings
back to dates. Not sure if this is the best place to do this
(could also be in the Calendar component itself, for example).