Make sure pre-populated dates stay on Calendar
Pre-populated dates that were being passed back to the Calendar
were beign initiated properly, but there was a bug where clicking
on a new day would cause the pre-populated days to disappear.

This is because the 'this.state.selectedDays' would be empty,
because the dates passed into the calendar wouldn't actually
update the state anywhere.

So now, if 'this.props.input.value' has values but the state
is empty, let's update the state from the input.

Note that I added an 'await' in front of setState because it's an
async function.