CSS tweaks for Summary component
I'm a bit on the fence about this whole thing because in practice
we have tabular data but tables for mobile are bad.
That means that for visual interfaces, our fake table is better
but for screenreaders a real table would make more sense.

This should be reconsidered later.

Things I tweaked:
- the "change" links were getting pushed out of the summary "table"
  so now they're not
- added a fatter border because it's more whimsical
- more consistently applied spacing rules
  - lately, I'm preferring margins to paddings because padding
    implies that the extra space lives with the element, whereas
    the majority of the time we just want things to be further away    from each other.