Add 'Remove date' button for dates in the date list
As shown in the mock-up, users will be able to remove
selected dates by clickingt a 'remove date' button beside the
list of dates that they've selected.

There was some funny behaviour happening when pressing spacebar
on a button whereby a (synthetic) click event was getting triggered
immediately on the subsequent button.
The automatic click event would have no x or y co-ordinates, and
its 'detail' (ie, number of times the mouse was clicked) would be
set to 0.
Therefore, we are ignoring click events where the detail === 0.
- https://www.w3schools.com/jsref/event_detail.asp
- https://stackoverflow.com/a/35381564/9728185

Also, the arrow-function-returning-an-arrow-function pattern was
a refactor I went on after reading a good medium article.
- https://medium.com/@machnicki/handle-events-in-react-with-arrow-functions-ede88184bbb

Final point:
Refactored Calendar tests so that we're not matching on the
string "Remove date" all the time.