Catch errors more comprehensively
Previously, our Submission component was catching GraphQL errors,
like an email being sent with the wrong fields, but wasn't
catching regular Errors, like a NetworkError.

Now it catches both.

I've left in a placeholder method that ignores the error, but it
would be the right place to log these things.