Add TelLink component
Telephone links are crap on desktops, but they seem like they're
pretty reasonable to use on phones.

Assuming that smaller screens are phones, I've added a component
that switches between an `a` tag with a "tel" href and just a normal
`span` on larger devices. Pretty nifty.

Also did moved all the react imports to the top of our component
test files just for some sanity.