Use passed-in value instead of internal state
Previously, I was using this.state.selectedDays
to store all the dates we had selected, and then
pushing it to this.props.input.value every so often.

This was a bad compromise as there then became two
places where the list of dates could be looked for
and referenced.

The issue I was having before was that at the end
of handleDayClick, the calendar doesn't automatically
re-render itself.

Setting internal state forces a re-render, which meant
that the visual behaviour made more sense, which is
why I kept the internal state around for so long.

The manual way to force a re-render is to use
this.forceUpdate(), but since we are setting state
based on whether we need an error message or not,
we don't have to use this.