Add list of days in right-hand column.
Also restyle calendar.

There is still a lot of polish needed to harden the
interaction with the calendar, but this is a reasonable
start.

Added focus styles and hover styles to the calendar
so that it can be navigated with a keyboard now.

Also updated tests to deal with new markup.