Add Google Analytics to track pageviews
Adds Google Analytics to every page using the
[`react-ga`](https://github.com/react-ga/react-ga) library.

Added some logic to make sure that Google Analytics is
only started when
- the `NODE_ENV` environment variable is set to 'production'
- the `RAZZLE_GA_ID` environment variable exists
  (this should be our Google tracking ID)

Otherwise, the app will run as normal and nothing will
be logged.

Currently there is a print statement each time a page
is logged by Google Analytics.
We should shut this off before we are in live beta or
whatever, but for now this seems okay.

//

This is the same library that the VAC team are using, so
we can help each other if need be.

Relevant VAC pull request:
https://github.com/cds-snc/vac-benefits-directory/pull/42

Also this example was helpful:
https://github.com/zeit/next.js/tree/canary/examples/with-react-ga