Upgraded all of the packages except razzle
It looks like when we run "yarn upgrade razzle", one of the
underlying dependencies that is upgraded breaks our hot reloading.

To be clear, doing a "yarn build" + "yarn start" is still working
fine. This is just for now to make development of the app easier.

And who knows, it might fix itself in a few days 🤞