Add email field to Registration form
Since our GraphQL API now expects an email field to be sent through
(this is something we are asking for from our users), we need
to include this field in our frontend otherwise all of our requests
to send emails will fail.

There was also a GraphQL validation error we ran into whereby the
email is being validated as a String in the local cache but it
was expecting a GraphQLEmail in the API. To solve this, we just
loosened up our API to accept a string and now it's all working.