Tiny config change for the API
Our inline-css package wants a URL from us so that it can give
us good links to stuff once we render everything to text.

But if we don't need to worry about this (ie, currently we don't
have any links in our emails so no problemo), then their advice
is to pass in an empty string.

BUT! Passing in an empty string would fail our falsey check in
api/index.js so now it doesn't.

Docs:
https://github.com/jonkemp/inline-css#optionsurl