Fix locale icu issues
So with Paul's help we were able to determine the reason the tests were failing in circle ci but not on my pc. There is a repo of locale names called an icu and node only stores a select few of them to keep its file size low, my pc for some reason had all the locales but other machines may not. So we add the full-icu module and append it to the test command.