Add React context components
I've been cooking these up in the lab all week.

The basic idea here is that we want to support a no-JS flow
so we need to:
- save values and access them across the app
- do this whether or not javascript is enabled
- be able to validate our fields on server and client side

Apollo link state is currently how we save things internally
but it's only saving values on the client so we can't access
any of our saved values from the server.

Instead, we're going to use React context to pass around our
state throughout our app, and we're going to save it as
an encryped cookie each time we update our context.

That way, whether we do a hard reload or an on-page ajax-y
request, we get access to the values we saved previously.

For more info / to check out the proof of concept driving this
forward, check out:
https://github.com/pcraig3/after-no-js