Apollo link state is gone-zo from Registration page
Removed `withApollo` from the registration page in favour of
`withProvider(withContext())`.

We are saving values to an internal store and also as browser
cookies. Wasn't as huge a rewrite as I expected although we
now have a console error because we're not passing the client
through to the header and footer components.