Remove most of Apollo link state config
We're still using Apollo link state for the language switching
part of the app, but the rest of the config can safely be removed.

Once we remove the language thing, it's possible we can get
rid of the Apollo provider entirely.

Also removed LinkStateDisplay because it was only useful while
we had everything saved in Apollo link state.