Render the Contact phone number first on error page
@zakkain was saying that sometimes the phone number is the better
contact information to privilege.

Quote:
This is the Error page. Right now it includes the same “contact”
module that we have on the final confirmation page (contact us,
here’s the email and phone number). Email is first, phone second.

For the error page though, we have a different context and user need.
Something went wrong and they didn’t complete their task - so I
think offering them an email option is a bit too slow
(especially if it’s actually the webform).

I’d like to prioritize the call centre phone number here, since
they can immediately call if they hit this error page and can be
rescheduled by CSC staff. And offer the webform as a secondary option.