Add dummy radio button / checkbox to our pages
This is nuts.

Pseudo element styles are not being applied correctly to the first
radio button and checkbox on the page, when in No-JS mode.

I don't know why this is happening really, but I don't want to
debug it at this point.

Adding a dummy radio to the registration page and a dummy checkbox
to the No-JS calendar page and then wrapping them in a hidden div.
- 'display: none' hides elements from visual interfaces and AT
- they are outside of our forms so they can't be submitted