Wrap all pages in I18nProvider
So here's the deal

Before, we had only a handful of pages that needed to be wrapped
with the React context stuff because it was only relevant for
pages that needed data.

So, for example, the registration page needed to populate itself
with previously submitted values, but the 404 page was just static
text.
At that point, it was just a matter of wrapping the pages that
needed it with `withProvider(withContext(PageName))`.

Now the baseline has shifted.

Since the <I18nProvider> needs an `en` or `fr` string, and since
that's only really available in `WithProvider`, anything that gets
a React.Provider will also be wrapped in an `I18nProvider`.

This means all pages now need to be wrapped with the
React context Provider since they all need to be translated.

As such, I just made that the default for all pages in `routes.js`
instead of copy-pasting it all into the various page components.

//

Deleted the RegistrationPage test because it was pointless.