Fix tests for FederalBanner after withContext
This is the general pattern we're going to see as we add withContext
to components.
In short, we have to export a wrapped version of the component
as well as a basic version of it for testing.

In our tests, we can pass in the hardcoded context object that
the PropTypes expect and we're all good.

//

Worth pointing out that there's an enzyme problem whereby the new
React Provider and Consumer nodes aren't recognized as legitimate
nodes, so any tests that rely on them will fail.
This should be addressed in the near future but for right now
we can't do a lot about it.
Source:
- https://github.com/airbnb/enzyme/issues/1509

Since the LanguageSwitcher is part of the FederalBanner (and
the LanguageSwitcher is wrapped in withContext), this means that
we can't `mount` our FederalBanner component until enzyme is updated.