NoJS Calendar error message shows up when it's meant to
Previously it was showing up on every pageload where the values
hadn't been saved yet.

This meant that the first time you ever loaded the page, it would
show up.

What we want is for the error to show up only after someone has
pressed "submit", so now that's what it does.

Also used the ErrorList just so that our errors are standardized.