Save dates from the JS calendar to the store as ISO date strings
Previously, the array of date objects was being saved to the store.
There was some kind of casting-to-a-string taking place but we
weren't doing it deliberately.

This way, we're a bit more explicit about what we want.

//

Now that we're doing this, the JS and No-JS calendars are saving
values to the store the same way. (yay consistency!)

Not that anyone will notice, but when toggling javascript,
saved dates will populate in either interface.