Preserve language when clearing data
Since we are saving _all_ data to the same cookie,
when the cookie is deleted, we lose the preferred language.

In order to prevent this, what I am doing is
- getting the language value from the cookie before it is deleted
- deleting the cookie
- redirecting to the 'cancel' page with a `language` query parameter
- the language will then be reset before the user gets the HTML
  response for the cancel page

//

This means that, unlike previously, cookies are never entirely
removed. Even if there was no cookie saved previously, (the
language would default to 'en'), cancelling will explicitly
set your language to english.