Translate footer links
This is basically a cut + paste from @Amirroracle's pull request.

It's great that this is working, but it's super gross.

The only other way I could think to do this was very similar,
passing an i18n object down through the context, but that seemed
just as complicated.

This split between in-page content and static text is fundamentally
unhelpful.

Ultimately, this might be a reason to give up lingui.

(which is a shame, because other than this one problem it's really
nice to use).