"Avis and Conditions" → "Avis"
Because we are hiding part of "Terms and Conditions", we had to
split up the <Trans> tags to accomodate for this.

But in french we don't have a translation for "and Conditions",
so I didn't fill one out.

This meant that, when switching from English to French, we would
see "Avis and Conditions".

Really, what we want is to hide the second half entirely when
french is selected. We don't need to bother about translating it.