Wrap No-JS calendar with a <span> to fix layout bug
This took a while to debug.

Because we are loading two different interfaces between the server
and the client, react is doing some clever diff thing so that it can
update everything super quick like.

But since there's a bunch of divs being nested, some of which
are undifferentiated, React can't figure out which parts of the
page to leave alone vs render anew.

When I took out the <div> wrapping the NoJSCalendar in CalendarPage
in commit: a051e58, I created a subtle layout bug that we didn't
notice on the JS calendar where the selected day list would always
render underneath and the error message would be rendered beside
the JS calendar.

Wrapping the noJS calendar in any element other than a <div> solves
this problem for us.

Kind of a weird one.