Fixed submit focus jumping
In JS-mode, the focus was being correctly applied to the error
list component, but we were relying on the browser to actually
scroll that element into view.

Chrome and Firefox were pretty good about it, but safari wasn't
scrolling our error element into view (I think because more
content was being added to the page or something).

Spent some time digging through @svinkle's tv-db app, and found
he was explicitly calling `window.scrollTo` in addition to setting
the focus on the element.

So we just copied him.

Source:
- https://github.com/svinkle/tv-db/blob/3502ae8cc7431079d888714f8a9aa485725d8b02/src/pages/Results.js#L77-L78