Remove day button alt text works in english and french
Both the "remove day" text and the date itself will be translated
when the app is switched into french.

Note that the alt-text for the images isn't having the date
added to it since there's pretty much no way to focus on the image
instead of the button.

Adding the alt-text means that it will pass the automated accessbility
scanning software audits (running axe, for example), but the aria-label
of the button is the thing screenreaders will end up on.