Switch to mailto link if sentry doesn't know about any errors
Was throwing an error before when window.Raven was undefined, so
first we're checking if Raven is truthy.

We're also looking for whether `Raven.lastEventId()` returns non-null
because `Raven.showReportDialog()` won't return anything if there is
no last event ID. (you can try this in the dev tools when sentry is
running).

Note:
Since we're doing a full page reload now (ie redirecting and then
reloading the app), Sentry isn't actually configured when the error
is thrown, so as of right now these errors are never caught.

ie, If an email fails to send, Sentry isn't going to catch it and
we're always going to give the user back the link to email us.

Note2:
The email I've used is the same as in the "Contact" link in the
footer of the page. Maybe we want to change it one day, and that's
fine with me.