Updated tests as they were failing given a start date where not enough dates would be available in the displayed month.
Tests were trying to click on indexes that they didn't exist in this case i.e. showing August and only Aug 30th was an available date to be clicked on.

The test will now count the available days in the current displayed month and click to the next month if not enough dates are available for the tests.

The days variable also needed updating to reflect the month swap happening as the days by default would be based on the current date, which would no longer match the calendar after the month swap.