Add Title component
Adding page titles to our various pages isn't particularly difficult,
but a more straightforward approach would mean that we have to
wrap all of our pages with `withI18n` just so that we can
translate ONE STRING.

Since this is clearly untenable, I'm creating a `<Title>`
component for each page to import.

The next problem we would run into is that lingui will only translate
literal strings, not variables, so all of the page titles have to
exist in the Title component so that our `extract` + `compile` steps
will pick them up.

Since page titles are static, this is actually a reasonable thing
for us to do.

This way, each page need only pass in its relative path to get its
title.