Update calendar tests to use valid days arrays for tests (#288)
* wip

* Update calendar tests to use valid days arrays for tests

This updates adds a new function to seek out a month to use for the tests that has at least 3 valid dates.

The tests have been fragile as we have been relying either upon the inital view to have multiple dates available or simulating clicks on the view to get to a month where we have enough dates.  This was falling out of sync with the dates trying to be selected.