Remove "request appointment" title from most pages
Our accessibility audits are mad at us for having something
that looks visually like a page title but isn't one.

After talking with @zakkain, we decided to remove the
"Request a new citizenship appointment" title from all pages
except the first page, and on the first page we're making it
an H1 because it's useful information you actually want to read.

That means that most pages now will just have the "skinny" blue
banner with the beta tag in it.