Add ingore rule for Span.Name
This seems to be a non issue for our app.  Case where this was failing is in the in Calendar where the next and back buttons are span elements with a role of button.  They do have an aria-label.

Further information can be found here:

https://www.w3.org/WAI/WCAG21/quickref/?versions=2.0#qr-ensure-compat-rsv

If custom controls are created, however, or interface elements are programmed (in code or script) to have a different role and/or function than usual, then additional measures need to be taken to ensure that the controls provide important information to assistive technologies and allow themselves to be controlled by assistive technologies.