Fixes issue where Calendar error max days ends up behind by a click.
The previous fix for this used forced render which fixed the error issue but, created an issue where users could not unselect dates after they we're saved to the cookie.

In this version we're passing the selected stated back to the force render and than updating state (using either the saved or state for the Calendar initial values).

This allows the Calendar to do a full re-render while maintaing the correct dates.