Try/catch when looking for git commit hash
When trying to boot our app outside of a git repo, calling
`execSync` with a git command was throwing an error.

Set up a try/catch so that errors won't kill the app, but then
was still getting an error message logged that said something like:
"fatal: not in a git repository".

Set the `stderr` output to `null` so that gitHash just returns
7 '0's without complaining if run on outside of a git repo.

Note: this is because when we are deploying to heroku, we
weren't in a git repo.