Set port dynamically if deploying to Heroku
Heroku prohibits you from setting a port on startup.
They are going to generate a port for you and then you are going
to listen for it as an env var.

- https://help.heroku.com/P1AVPANS/why-is-my-node-js-app-crashing-with-an-r10-error

Unfortunately, Razzle defaults to setting port 3000 if you
don't set one yourself.

- https://github.com/jaredpalmer/razzle#build-time-variables

There's some issues about this on the razzle repo, and ultimately
this is where I found the solution.

Source: https://github.com/jaredpalmer/razzle/issues/356#issuecomment-366275253