Add new test going through the service passing query params
This is a feature we support but we haven't had any tests
for it until now.

Submitting query parameters for the form page is pretty easy
to test, but on the calendar page it gets a bit weird.

- Since cypress is always running with JS in the client,
  I'm not trying to turn it off or anything
- Since our JS calendar page won't show errors when query
  parameters are passed through, I'm not testing for that
- Since our calendar logic doesn't actually validate your
  dates are within a certain range, I'm hardcoding some
  invalid dates

The test therefore tests that calendar days _can_ be passed
through as query parameters and they _will_ show up on the
review page.

It's *not* testing that we're passing valid dates through.