Load Raven Async
This update modifies the Raven install to make it load async

When recording load times in Google Chrome the initial costly events are caused by the Raven install.  It's preferable to offset the install in order to get to HTML parsing sooner.

By moving to async loading we can continue to capture error logs while boosting the amount of time it takes to get to the HTML parsing.

Details:
https://docs.sentry.io/clients/javascript/install/#async-loading