Boost homepage load time using cache
The loading times for slow 3G connections has been quite slow 30+secs.  This in part is due to the bundle size and parse time for the bundle.

Given the the interactivity level and content for the initial landing page we can boost load time but serving up a cached version of the landing page and skip loading the client bundle for the initial page.