Translation [Month] language toggle issue
When toggling the language the month name was not getting translated for the text Citizenship appointments in [MONTH] are scheduled on Tuesdays and Wednesdays.

If you used the next month button the issue corrects itself.

This update adds componentDidUpdate to check for a change in the locale prop.

Cypress tests have been added.