Add Sentry notification for missing data redirects
Routing users back to fill in missing data shouldn't happen given a typical user flow.  If we end up flagging missing data and redirecting we should be notified we can look further into potential issues.

This update sends a simple message to Sentry for these case so the information is logged and we will receive an email notification.