Fix fonts preloading + remove theme.weight
The way we were adding fonts before wasn't actually using our
preloaded fonts.
We were preloading our fonts, but then redownloading them later.

Instead, I used this example:
https://github.com/zachleat/web-font-loading-recipes/blob/master/preload.html

The other thing that changed is that I am assigning both weights
to the same font family, but giving them different weights.

This means that when we want to switch between fonts in our app,
all we have to do is switch the weights.

This solves a couple of problems for us:
- weights now work the same as on our fallback (system) font
- means our font renders properly across differnet browsers
  (ie, this fixes our weird kerning problem)

Since we can just set font weights to 700 and be done with it,
I've removed the `theme.weight` object entirely and deleted
everywhere it was being used in the app.

Since header elements default to `font-weight: 700`, I didn't
have to explicitly define it in most cases.

Also:
- Removed `strong` declaration that was redundant
- Removed Bold element that wasn't being used