Add a text prompt signaling user needs to remove days
Users were still being confused when navigating back to the
Calendar Page.
It is apparent that removing days is not as obvious as we
hoped at this point in the user flow.

Adding a text prompt for the user to remove days when navigating
back to the calendar after having already saved dates.

A daysModified variable has been added to Calendar.js that is set
to true whenever a selected day is clicked.
This will be used in conjunction with threeDatesArePicked as a
switch/trigger for this message.