Prevent IE11 from focusing on Canada government logo
In our last round of accessibility testing, there was a bug uncovered
where IE11 would automatically focus on our Canadian Federal logo
(ie, an svg), even though this is a non-interactive element
and does not have a focus indicator.

Fix is to add a focusable="false" attribute to the svg.
- https://github.com/PolymerElements/iron-iconset-svg/issues/46

Went into GCUI to make the change, and added tests for it in there.
- https://github.com/cds-snc/gcui/pull/23