Update test to use param that isn't in use
Given a case where ?not-valid=true

This PR includes a case the code is triggering a submit event (which seems to be the proper way Final Form Suggests)

This in turn shows error messages if fields are missing automatically when the page is loaded.

The now updated test was failing as all the fields are filled in via the test.

In which case when visiting the page with not-valid isn't correct

- Submit happens and the data is valid
- We end up on the Calendar page
- We look for fullName that doesn't exist
- Test was failing