Initial Subdomain setup
For local testing use vcap.me:3004 i.e. http://mtrl.vcap.me:3004

Sets up the server to handle locations via subdomain.

Valid subdomains have been whitelisted.

If a subdomain is found but is not on the list it will redirect.

If no subdomain is found for now it will default to Vancouver.

redirect to generic Not Found page