1. Adds ability to have blocked days on the calendar
This adds a 'blocked' property to the location object to block off calendar days

export const test_location = {
  recurring: {
    ...
  },
  blocked: '2018-10-02, 2018-10-03, 2018-11-21', // use CSV format => 2018-10-02, 2018-10-03, 2018-11-21
}

If the property exists the CSV string will be parsed and dates blocked on those values

2. Adds utility method to parse a CSV string of dates

dateSetFromString takes a 'user entered' date string  i.e. 2018-10-17, 2018-11-06, 2018-11-07
and creates a new JavaScript set.

This will be used for checking if a given date is in the list set.has('2018-10-17')