Add location to app context
This commit location importing into the app.
- tries to `require` the location file based on a passed-in string
- catches errors (ie, location file doesn't exist) and sets a
  placeholder object with an id of the passed-in string

Once we have the location (or not), we cache it and add it to the
app context.

Since our Context.Provider has been so-far wrapping Page components,
it meant that it would be reset whenever there was a change to a new
page. So, going from the register page to the calendar page would
clear the previous context.

The reason we weren't losing data between pages is that we re-hydrate
our context from the cookies every time.

However, with the location, we can set it on the server (from any page,
since we're using subdomains to pick the location), and then pass it
into the app from there.
Once you're in SPA-land, you just have the location cached in the
context and we want to keep it around as you go between pages.

This was looking like it would be hard to do until (🎉🎉), it
turns out that Providers can be arbitrarily nested.

Example:
- https://codesandbox.io/s/lrvv8w784q

Also, After passes whatever data gets retuned from `getInitialProps`
to itself in `Document.js` (run on the server), and `client.js`
(run on the client).

In practice, this means:
- we can get location in getInitialProps the first time the site is loaded
- it gets run once on the server, so we pass it through to WithProvider
  constructor
- when it's run on the client, we can pass it through to a Context.Provider
  wrapping After in client.js
  - this last step is only run the first time the JS parses, but the location
    will remain in the context until the JS is interrupted (ie, a new pageload
    or something)
- If there is a new pageload, we can re-load the location server-side.

It's a pretty good solution, it feels like.

Location is lodaed dynamically on the server the first time it's needed and
then passed through to the client, where it's cached until it changes or we
get another serverside request.