Use getGlobalLocation to get location
Passing the context through the context is great, but a lot of
our date-calculating code needs to know about the location outside
of React-land, so we need to be able to serve up the location
somewhere else.

Wrote a small 'set once and cache' JS module (using the module
pattern: https://toddmotto.com/mastering-the-module-pattern/)
that will save a location once and then return the saved
location thereafter.

Works perfectly on the server, but since the location object we're
saving gets lost between the server and the client, we have to
set this once on the client as well (on first render).

After the first render, it will be cached on the client, so it
doesn't need to be re-set as we move between pages.

To set it on the client, we're checking if windowExists and if
there's no subdomain set (ie, it's only being set on the server).